Vim's like behavior when opening and saving files
If an arguemnt passed from the command line (name of the current working file)
doesn't exist at that specific path, a temporary file will be
created until user decides to save it or close it. Depending on the users
action, the temporary file will be renamed to the original file name
or will be removed.
Error handling on opening files are now more verbose, providing details
based on the different situations that could arise.